requirements.txt
google-adk
fastapi
uvicorn
google-cloud-firestore
python-dotenv